[CH847] Api key strategy (#18)
* Add API Key Strategy

* Merge changes from dev

* Remove API Key helpers

* Prune packages

* fix publish

* fix publish

* Return venue and org
