+++ lib/index.js
@@ -1,7 +1,8 @@
 /**
  * Module dependencies.
  */
-var Strategy = require('./bypass_strategy');
+const Strategy = require('./bypass_strategy');
+const KeyStrategy = require('./key_strategy');
 
 /**
  * Expose `Strategy` directly from package.
@@ -12,4 +13,5 @@ exports = module.exports = Strategy;
  * Export constructors.
  */
 exports.Strategy = Strategy;
+exports.KeyStrategy = KeyStrategy;
 exports.Auth = require('./bypass_auth');
